+++ RELEASE_NOTES.md
@@ -38,6 +38,11 @@ construct.
 character atoms. This is necessary to ensure that we can always reconstruct
 the entity identifier from the internal entity prefix.
 
+* FIXED: Bug when converting an internal entity prefix to the entity
+identifier when the identifier includes the character used as delimiter
+between the entity functor from the entity number of parameters in the
+entity internal prefix.
+
 * FIXED: Cases where sending a database message to an invalid object
 identifier would result in an existence error instead of a type error.